Fix failing test by moving url utilities to urlUtils.ts (BL-8874)
